requirements.txt
openpyxl
pandas
threaded
speedtest-cli